Handler error exception Implementation

+++ src/main/java/com/carsil/userapi/service/UserService.java
@@ -3,10 +3,14 @@ package com.carsil.userapi.service;
 import com.carsil.userapi.model.User;
 import com.carsil.userapi.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -17,33 +21,48 @@ public class UserService {
     @Autowired
     private PasswordEncoder passwordEncoder;
 
-
     public List<User> getAll() {
         return userRepository.findAll();
     }
 
+
     public User create(User u) {
+        Objects.requireNonNull(u, "El usuario no puede ser nulo.");
+        if (u.getPassword() == null || u.getPassword().isBlank())
+            throw new IllegalArgumentException("La contraseña es obligatoria y no puede estar vacía.");
         u.setPassword(passwordEncoder.encode(u.getPassword()));
         return userRepository.save(u);
     }
 
     public void delete(Long id) {
+        if (!userRepository.existsById(id))
+            throw new NoSuchElementException("Usuario no encontrado con id: " + id);
         userRepository.deleteById(id);
     }
 
     public boolean validateLogin(String name, String rawPassword) {
-        Optional<User> user = userRepository.findByName(name);
-        return user.isPresent() &&
-                passwordEncoder.matches(rawPassword, user.get().getPassword());
+
+        if (!StringUtils.hasText(name) || !StringUtils.hasText(rawPassword)) {
+            throw new IllegalArgumentException("Usuario y contraseña son obligatorios");
+        }
+        var user = userRepository.findByName(name)
+                .orElseThrow(() -> new BadCredentialsException("Credenciales inválidas"));
+
+        if (!passwordEncoder.matches(rawPassword, user.getPassword())) {
+            throw new BadCredentialsException("Credenciales inválidas");
+        }
+        return true;
     }
 
     public Optional<User> getById(Long id) {
-        return userRepository.findById(id);
+        return Optional.ofNullable(userRepository.findById(id).
+                orElseThrow(() -> new NoSuchElementException(
+                        "Usuario con ID " + id + " no fue encontrado.")));
     }
 
     public Optional<User> getByName(String name) {
-        return userRepository.findByName(name);
+        return Optional.ofNullable(userRepository.findByName(name).orElseThrow(() -> new NoSuchElementException(
+                "No existe un usuario registrado con el nombre '" + name + "'.")));
     }
-
 }
